JDK21: Key Encapsulation Mechanism API: Refactor Kyber and Security Classes
